SPRM.py: add default values to `main` function
These are handled by `argparse` when SPRM is run through the command line,
but the CODEX pipeline imports and calls `SPRM.main` in Python code. We
shouldn't know or care about the full path to the SPRM options file in the
CODEX pipeline, so set a default value in the `main` function definition.