Save pixel arrays at full resolution
Using the matplotlib `plt.imshow` method isn't appropriate for this; that
method is intended for image plots with titles and axis labels and such, and
resamples the image to fit in the plot region.

Since the full-resolution pixel data is available as NumPy arrays, build
and save images with Pillow. The matplotlib color mapping functionality is
still used to map continuous values to the viridis color palette when
appropriate.

3-channel PCA data is saved directly, without any matplotlib color manipulation.