Use sparse storage for adjacency matrix (in memory and on disk)
Storing this as a dense matrix is a significant waste of memory, disk space,
and CPU time to create/store/load this matrix, since the number of nonzero
entries in the matrix will scale (more or less) linearly with the number
of cells.

Change the in-memory representation to a scipy.sparse.dok_matrix, which
behaves identically but avoids allocating a large block of memory that scales
quadratically with the number of cells. When writing to disk, convert to CSR
format (sorting by row, then by column index) and write to disk in the MTX
file format. For a *small* test dataset with 11616 cells, the dense CSV
representation is 515MB on disk, and the MTX version is orders of magnitude
smaller while containing the same information.